+++ tekore/_model/chapter/full.py
@@ -1,0 +1,20 @@
+from typing import Optional
+from dataclasses import dataclass
+
+from .base import Chapter
+from ..member import Restrictions
+from ..audiobook.full import FullAudiobook
+
+
+@dataclass(repr=False)
+class FullChapter(Chapter):
+    """Complete chapter object."""
+
+    audiobook: FullAudiobook
+    restriction: Optional[Restrictions] = None
+
+    def __post_init__(self):
+        super().__post_init__()
+        self.audiobook = FullAudiobook(**self.audiobook)
+        if self.restriction:
+            self.restriction = Restrictions(**self.restriction)